Deleted Box Struct
Due to memory space, a game board could not be larger than a 4x4, since each board space (box) was a struct. The logic was changed so that all the box information is instead stored as unsigned ints, where the first 4 values indicate the up, down, left, and right walls. And the 5th value indicates whether the ball is in that space or not. This allows for nearly an infinitely large board size! A 10x10 board was hardcoded in and works as expected!